+++ HansRoslinger/client/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Meteor } from 'meteor/meteor';
-import { App } from '/imports/ui/App';
+import App from '/imports/ui/App';
 
 Meteor.startup(() => {
   const container = document.getElementById('react-target');
