+++ HansRoslinger/imports/ui/pages/SettingsPage.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Settings } from '../Settings';
+
+export default function SettingsPage() {
+  return (
+    <div>
+        <Settings/>
+    </div>
+  );
+}
