+++ HansRoslinger/package.json
@@ -23,6 +23,8 @@
     "react": "^18.2.0",
     "react-chartjs-2": "^5.3.0",
     "react-dom": "^18.2.0",
+    "react-router": "^6.14.1",
+    "react-router-dom": "^6.14.1",
     "react-webcam": "^7.2.0",
     "recharts": "^2.15.2",
     "tailwindcss": "^4.1.6"
